requirements.txt
streamlit==1.29.0
google-generativeai==0.3.2
speechrecognition==3.10.0
gtts==2.4.0
pydub==0.25.1
pandas==2.1.4
python-dotenv==1.0.0
Pillow==10.1.0
plotly==5.18.0
streamlit-chat==0.1.1
